MAGECLOUD-2901: [Spike] Prefixes for DB Tables are not Considered (#478)
* MAGECLOUD-2901: [Spike] Prefixes for DB Tables are not Considered
