+++ src/Command/ConfigDump/Generate.php
@@ -143,7 +143,10 @@ class Generate
          * Adding locales for admin user.
          */
         $newConfig['admin_user']['locale']['code'] = array_column(
-            $this->connection->select('SELECT DISTINCT `interface_locale` FROM `admin_user`'),
+            $this->connection->select(sprintf(
+                'SELECT DISTINCT `%s` FROM `admin_user`',
+                $this->connection->getTableName('interface_locale')
+            )),
             'interface_locale'
         );
