+++ src/Test/Unit/DB/ConnectionTest.php
@@ -5,6 +5,7 @@
  */
 namespace Magento\MagentoCloud\Test\Unit\DB;
 
+use Magento\MagentoCloud\Config\Database\MergedConfig;
 use Magento\MagentoCloud\DB\Connection;
 use Magento\MagentoCloud\DB\Data\ConnectionFactory;
 use Magento\MagentoCloud\DB\Data\ConnectionInterface;
@@ -39,9 +40,14 @@ class ConnectionTest extends TestCase
     private $loggerMock;
 
     /**
-     * @var ConnectionInterface
+     * @var ConnectionInterface|MockObject
      */
     private $connectionDataMock;
+
+    /**
+     * @var MergedConfig|MockObject
+     */
+    private $mergedConfigMock;
 
     /**
      * {@inheritdoc}
@@ -52,6 +58,7 @@ class ConnectionTest extends TestCase
     {
         $this->pdoMock = $this->createMock(\PDO::class);
         $this->statementMock = $this->createMock(\PDOStatement::class);
+        $this->mergedConfigMock = $this->createMock(MergedConfig::class);
         $this->loggerMock = $this->getMockForAbstractClass(LoggerInterface::class);
         $this->connectionDataMock = $this->getMockForAbstractClass(ConnectionInterface::class);
 
@@ -66,7 +73,8 @@ class ConnectionTest extends TestCase
 
         $this->connection = new Connection(
             $this->loggerMock,
-            $connectionFactoryMock
+            $connectionFactoryMock,
+            $this->mergedConfigMock
         );
 
         $reflection = new \ReflectionClass(get_class($this->connection));
@@ -156,15 +164,6 @@ class ConnectionTest extends TestCase
         $this->connection->close();
     }
 
-    public function testCount()
-    {
-        $this->statementMock->expects($this->once())
-            ->method('rowCount')
-            ->willReturn(1);
-
-        $this->assertSame(1, $this->connection->count('SELECT 1'));
-    }
-
     public function testAffectingQuery()
     {
         $bindings = [
@@ -203,5 +202,44 @@ class ConnectionTest extends TestCase
             ->willReturn(true);
 
         $this->connection->query('SELECT 1', $bindings);
+    }
+
+    /**
+     * @param array $mergedConfig
+     * @param string $tableName
+     * @param string $expectedTableName
+     * @dataProvider getTableNameDataProvider
+     */
+    public function testGetTableName(array $mergedConfig, string $tableName, string $expectedTableName)
+    {
+        $this->mergedConfigMock->expects($this->once())
+            ->method('get')
+            ->willReturn($mergedConfig);
+
+        $this->assertEquals(
+            $expectedTableName,
+            $this->connection->getTableName($tableName)
+        );
+    }
+
+    /**
+     * @return array
+     */
+    public function getTableNameDataProvider(): array
+    {
+        return [
+            'empty prefix' => [
+                [],
+                'table',
+                'table',
+            ],
+            'non empty prefix' => [
+                [
+                    'table_prefix' => 'ece_',
+                ],
+                'table',
+                'ece_table',
+            ],
+        ];
     }
 }
